order-service/order_service/orders/views/views_v1.py
from rest_framework import viewsets, mixins
from rest_framework.permissions import AllowAny  
from rest_framework.response import Response
from rest_framework.decorators import api_view
from rest_framework import status
import logging

from utils.shopcart_client import shopcart_client
from utils.product_client import product_client
from utils.shop_client import shop_client
from ..models import * 
from ..serializers import *

logger = logging.getLogger(__name__)



#Order Create
@api_view(['GET', 'POST'])
def orders_list_create(request):
    if request.method == 'GET':
        orders = Order.objects.all().order_by('-id')
        serializer = OrderSerializer(orders, many=True)
        return Response(serializer.data)
    
    elif request.method == 'POST':
        serializer = OrderSerializer(data=request.data)
        if serializer.is_valid():
            serializer.save()
            return Response(serializer.data, status=status.HTTP_201_CREATED)
        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
    
    
@api_view(['GET', 'PATCH', 'DELETE'])
def orders_detail(request, pk):
    try:
        order = Order.objects.get(pk=pk)
    except Order.DoesNotExist:
        return Response({"error": "Order not found"}, status=status.HTTP_404_NOT_FOUND)

    if request.method == 'GET':
        serializer = OrderSerializer(order)
        return Response(serializer.data)

    elif request.method == 'PATCH':
        serializer = OrderSerializer(order, data=request.data, partial=True)
        if serializer.is_valid():
            serializer.save()
            return Response(serializer.data)
        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)

    elif request.method == 'DELETE':
        order.delete()
        return Response(status=status.HTTP_204_NO_CONTENT)



#OrderItem
# GET / POST
@api_view(['GET', 'POST'])
def orderitems_list_create(request):
    if request.method == 'GET':
        items = OrderItem.objects.select_related('order').all().order_by('-id')
        serializer = OrderItemSerializer(items, many=True)
        return Response(serializer.data)
    
    elif request.method == 'POST':
        serializer = OrderItemSerializer(data=request.data)
        if serializer.is_valid():
            serializer.save()
            return Response(serializer.data, status=status.HTTP_201_CREATED)
        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)


# GET / PATCH / DELETE /order-items/<id>/
@api_view(['GET', 'PATCH', 'DELETE'])
def orderitems_detail(request, pk):
    try:
        item = OrderItem.objects.get(pk=pk)
    except OrderItem.DoesNotExist:
        return Response({"error": "OrderItem not found"}, status=status.HTTP_404_NOT_FOUND)

    if request.method == 'GET':
        serializer = OrderItemSerializer(item)
        return Response(serializer.data)

    elif request.method == 'PATCH':
        serializer = OrderItemSerializer(item, data=request.data, partial=True)
        if serializer.is_valid():
            serializer.save()
            return Response(serializer.data)
        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)

    elif request.method == 'DELETE':
        item.delete()
        return Response(status=status.HTTP_204_NO_CONTENT)
    
@api_view(['POST'])
def create_order_from_shopcart(request):
    user_id = str(request.user.id)

    shopcart_data = shopcart_client.get_shopcart_data(user_id)

    if not shopcart_data:
        return Response({"detail": "Shopcart not found"}, status=status.HTTP_404_NOT_FOUND)

    items = shopcart_data.pop('items', [])
    order_data = {"user_id": user_id}

    order_serializer = OrderSerializer(data=order_data)
    if order_serializer.is_valid():
        order = order_serializer.save()
        logger.info(f'Order created successfully - Order ID: {order.id}, User: {user_id}, Items: {len(items)}')
    else:
        logger.error(f'Order serializer errors: {order_serializer.errors}')
        return Response(order_serializer.errors, status=status.HTTP_400_BAD_REQUEST)

    for item in items:
        order_item_data = {
            'order': order.id,  # bu sətri dəyişəcəyik
            'product_variation': item.get('product_variation'),
            'quantity': item.get('quantity', 1),
            'status': 1,  
            'price': 0  
        }

        # ✅ Əsas dəyişiklik: order instance göndəririk, id yox
        order_item_data['order'] = order

        item_serializer = OrderItemSerializer(data=order_item_data)
        if item_serializer.is_valid():
            item_serializer.save()
        else:
            logger.error(f'Order item serializer errors: {item_serializer.errors}')
            return Response(item_serializer.errors, status=status.HTTP_400_BAD_REQUEST)

    return Response(
        {"message": "Order and items created successfully"},
        status=status.HTTP_201_CREATED
    )


@api_view(['PATCH'])
def update_order_item_status(request, pk):
    try:
        item = OrderItem.objects.get(pk=pk)
    except OrderItem.DoesNotExist:
        return Response({"error": "OrderItem not found"}, status=status.HTTP_404_NOT_FOUND)
    
    new_status = request.data.get("status")
    if new_status not in dict(OrderItem.Status.choices):
        return Response({"error": "Invalid status"}, status=status.HTTP_400_BAD_REQUEST)

    variation_id = str(item.product_variation)
    variation_data = product_client.get_variation(variation_id)
    product_id = str(variation_data.get("product_id")) if variation_data else None
    product_data = product_client.get_product(product_id)
    shop_id = str(product_data.get("shop_id")) if product_data else None
    user_id = str(request.user.id)
    user_shop_ids = shop_client.get_user_shop_ids(user_id)

    if shop_id not in user_shop_ids:
        return Response({"error": "Forbidden: You do not own this shop's item"}, status=status.HTTP_403_FORBIDDEN)

    item.status = new_status
    if not item.product_id:
        item.product_id = product_id
    if not item.shop_id:
        item.shop_id = shop_id
    item.save()

    item.order.check_and_approve()

    serializer = OrderItemSerializer(item)
    return Response(serializer.data, status=status.HTTP_200_OK)