order-service/pyproject.toml
[project]
name = "order-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.7",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.28.0",
    "gunicorn>=23.0.0",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.4.2",
    "pytest-django>=4.11.1",
    "python-dotenv>=1.1.1",
    "whitenoise>=6.5",
    "psycopg[binary] (>=3.2.10,<4.0.0)",
    "drf-yasg>=1.21.7",
    "httpx>=0.28.1",
    "requests>=2.32.3",
]